+++ Backend/src/main/java/com/homiwood/peliculas/service/UsuarioService.java
@@ -1,10 +1,12 @@
 package com.homiwood.peliculas.service;
 
-import com.homiwood.peliculas.model.Usuario;
-import com.homiwood.peliculas.repository.UsuarioRepository;
-import org.springframework.stereotype.Service;
+import com.homiwood.peliculas.dto.CrearUsuarioRequest;
 import com.homiwood.peliculas.exception.DuplicateResourceException;
 import com.homiwood.peliculas.exception.NotFoundException;
+import com.homiwood.peliculas.model.Usuario;
+import com.homiwood.peliculas.repository.UsuarioRepository;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
 
@@ -12,24 +14,35 @@ import java.util.List;
 public class UsuarioService {
 
     private final UsuarioRepository usuarioRepository;
+    private final PasswordEncoder passwordEncoder;
 
-    public UsuarioService(UsuarioRepository usuarioRepository) {
+    public UsuarioService(
+            UsuarioRepository usuarioRepository,
+            PasswordEncoder passwordEncoder
+    ) {
         this.usuarioRepository = usuarioRepository;
+        this.passwordEncoder = passwordEncoder;
     }
 
     public List<Usuario> listarUsuarios() {
         return usuarioRepository.findAll();
     }
 
-    public Usuario crearUsuario(Usuario usuario) {
+    public Usuario crearUsuario(CrearUsuarioRequest request) {
 
-        if (usuarioRepository.existsByUsername(usuario.getUsername())) {
+        if (usuarioRepository.existsByUsername(request.getUsername())) {
             throw new DuplicateResourceException("El username ya está registrado");
         }
 
-        if (usuarioRepository.existsByEmail(usuario.getEmail())) {
+        if (usuarioRepository.existsByEmail(request.getEmail())) {
             throw new DuplicateResourceException("El email ya está registrado");
         }
+
+        Usuario usuario = new Usuario();
+        usuario.setNombre(request.getNombre());
+        usuario.setUsername(request.getUsername());
+        usuario.setEmail(request.getEmail());
+        usuario.setPasswordHash(passwordEncoder.encode(request.getPassword()));
 
         return usuarioRepository.save(usuario);
     }
@@ -38,7 +51,13 @@ public class UsuarioService {
         return usuarioRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException("Usuario no encontrado"));
     }
+
     public void eliminarUsuario(Long id) {
+
+        if (!usuarioRepository.existsById(id)) {
+            throw new NotFoundException("Usuario no encontrado");
+        }
+
         usuarioRepository.deleteById(id);
     }
 }
